+++ tests/testthat/test-oauth.R
@@ -53,3 +53,62 @@ with_mock_api({
     )
   })
 })
+
+with_mock_api({
+  test_that("we can retrieve the AWS viewer credentials", {
+    con <- Connect$new(server = "https://connect.example", api_key = "fake")
+    expect_true(validate_R6_class(con, "Connect"))
+    credentials <- get_aws_credentials(con, user_session_token = "user-session-token")
+    expect_equal(
+      credentials,
+      list(
+        accessKeyId = "aws_access_key_id",
+        secretAccessKey = "aws-secret-access-key",
+        sessionToken = "aws-session-token",
+        expiration = "2025-01-01T00:00:00Z"
+      )
+    )
+  })
+
+  test_that("we can retrieve the AWS content credentials with an explicit token", {
+    con <- Connect$new(server = "https://connect.example", api_key = "fake")
+    expect_true(validate_R6_class(con, "Connect"))
+    credentials <- get_aws_content_credentials(con, content_session_token = "content-session-token")
+    expect_equal(
+      credentials,
+      list(
+        accessKeyId = "aws_access_key_id",
+        secretAccessKey = "aws-secret-access-key",
+        sessionToken = "aws-session-token",
+        expiration = "2025-01-01T00:00:00Z"
+      )
+    )
+  })
+
+  test_that("we can retrieve the AWS content credentials with an env var", {
+    Sys.setenv(CONNECT_CONTENT_SESSION_TOKEN = "content-session-token")
+    on.exit(Sys.unsetenv("CONNECT_CONTENT_SESSION_TOKEN"))
+
+    con <- Connect$new(server = "https://connect.example", api_key = "fake")
+    expect_true(validate_R6_class(con, "Connect"))
+    credentials <- get_aws_content_credentials(con)
+    expect_equal(
+      credentials,
+      list(
+        accessKeyId = "aws_access_key_id",
+        secretAccessKey = "aws-secret-access-key",
+        sessionToken = "aws-session-token",
+        expiration = "2025-01-01T00:00:00Z"
+      )
+    )
+  })
+
+  test_that("we cannot retrieve the AWS content credentials without a token or env var", {
+    con <- Connect$new(server = "https://connect.example", api_key = "fake")
+    expect_true(validate_R6_class(con, "Connect"))
+    expect_error(
+      get_aws_content_credentials(con),
+      "Could not find the CONNECT_CONTENT_SESSION_TOKEN environment variable."
+    )
+  })
+})